+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/graph/IrBindingStack.kt
@@ -12,7 +12,6 @@ import dev.zacsweers.metro.compiler.decapitalizeUS
 import dev.zacsweers.metro.compiler.diagnostics.Note
 import dev.zacsweers.metro.compiler.expectAs
 import dev.zacsweers.metro.compiler.graph.BaseBindingStack
-import dev.zacsweers.metro.compiler.graph.BaseTypeKey
 import dev.zacsweers.metro.compiler.ir.ClassFactory
 import dev.zacsweers.metro.compiler.ir.IrContextualTypeKey
 import dev.zacsweers.metro.compiler.ir.IrTypeKey
@@ -352,20 +351,6 @@ internal interface IrBindingStack :
 
     fun empty() = EMPTY
   }
-}
-
-internal inline fun <
-  T,
-  Type : Any,
-  TypeKey : BaseTypeKey<Type, *, *>,
-  Entry : BaseBindingStack.BaseEntry<Type, TypeKey, *>,
-  Impl : BaseBindingStack<*, Type, TypeKey, Entry, Impl>,
-> Impl.withEntry(entry: Entry?, block: () -> T): T {
-  if (entry == null) return block()
-  push(entry)
-  val result = block()
-  pop()
-  return result
 }
 
 internal fun Appendable.appendBindingStackEntries(